Change values to byte instead of floats

+++ reactive-ev3-server/src/main/java/de/herbstcampus/DataSampler.java
@@ -3,5 +3,5 @@ package de.herbstcampus;
 import reactor.core.publisher.Flux;
 
 interface DataSampler {
-  Flux<float[]> sample(long sampleRate);
+  Flux<byte[]> sample(long sampleRate);
 }
